test sampling gamma
