+++ src/main/java/rapaio/core/distributions/Poisson.java
@@ -33,7 +33,7 @@ import rapaio.sys.WS;
  * if these events occur with a known average rate and independent
  * of the last occurrence of last events.
  */
-public class Poisson implements Distribution {
+public class Poisson extends AbstractDistribution {
 
     private static final long serialVersionUID = 2013039227493064895L;
     private final double lambda;
@@ -70,40 +70,6 @@ public class Poisson implements Distribution {
         if (x < 0)
             return 0.0;
         return MTools.incompleteGammaComplement(Math.floor(x + 1), lambda);
-    }
-
-    @Override
-    public double quantile(double p) {
-        if (p == 1)
-            return Double.POSITIVE_INFINITY;
-
-        double cdf0 = cdf(0);
-        if (p <= cdf0)
-            return 0;
-
-        // unbounded binary search
-        int low = 0;
-        int up = 1;
-
-        // double up until we found a bound
-        double cdf_up = cdf(up);
-        while (cdf_up <= p) {
-            up *= 2;
-            cdf_up = cdf(up);
-        }
-        while (low < up) {
-            int mid = Math.floorDiv(low + up, 2);
-            if (mid == low)
-                return up;
-            double cdf_mid = cdf(mid);
-            if (cdf_mid < p) {
-                low = mid;
-            } else {
-                up = mid;
-            }
-        }
-
-        return 0;
     }
 
     @Override
